requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.1
google-cloud-logging==2.7.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
mock==4.0.3
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
slack-sdk==3.13.0
smmap==5.0.0
terrasnek==0.1.7
time-machine==2.5.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
